src/requirements.txt
pytest==8.3.4
playwright==1.49.1
pytest-playwright==0.6.2
allure-pytest==2.13.5
python-dotenv==1.0.1
requests==2.32.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0